[TASK] Add Settings.yml, Includes.txt and more ...
We start by doing this: Add Settings.yml, add Includes.txt, use a
better 'conf.py' file. It is the "current up to date" generic one
with the "TYPO3 codeblock" at the end.

Add .. include:: Includes.txt in Documentation/Index.rst to show
what should be done in each ReST file. (Will do that later
programmatically, I think.)

Change all Images.txt files to not use the height and width
specification given there as those numbers reflect what OpenOffice
once thought was the correct size. But it isn't the actual size of
the screenshots but a scaled down number. We better ignore
those numbers.

This patch can serve as a model of what needs to be done with other
manuals as well.

Change-Id: Iec9374ab90c29b5d583fc2b08012da78026c9282
Resolves: #42322