[TASK] Eliminate "generated" headlines
In the OpenOffice document we sometimes left out one level of headlines.
During the migration to ReST, however every level of headlines, if it
existed or not, was written in the .rst files; also the missing ones.
These headlines now show up as "((generated))" in the files and in the
rendered output.

Replace each "generated" headline with the one which should actually be
there. 

Change-Id: Ib8883576e02d84d1629b069837e16169a4ba05c7
Resolves: #42573