[TASK] Improve pages "data types" and "imageLinkWrap"
Better now:

Data types now have a section (= headline) of its own. The headline
text is the name of the data type [1]. Crossreferences thus get the
correct linktext automatically. And the section improve the
intra-page navigation.

Some lists are rewritten as table now [2,3].

Introduced data type "stdWrap" that will need a description. [4]

The page "imageLinkWrap" [5] makes use of the data types and may serve
as a demo. This page is still under construction.

Change-Id: Ibd66a5ebdd07ab0f094f0bb24c456329a8ac73da
Resolves: #44298
Reviewed-on: https://review.typo3.org/17362
Reviewed-by: Martin Bless
Tested-by: Martin Bless