[TASK] Update documentation of stdWrap of cObject TEXT
The property ".stdWrap" is only mentioned generally as part of the
"stdWrap functions" while it in contrast is explicitly mentioned for the
other cObjects. Document it explicitly here as well!
Update the note ("very rootlevel") accordingly; discourage usage of
those stdWrap properties, which are located directly on the rootlevel of
the cObject.

Update the documentation to use stdWrap consistently. Update examples
through TSref to no longer use the rootlevel properties, but to either
use the stdWrap properties of ".value" or the property ".stdWrap".

Change-Id: I304b5e43caae29251a097e435ffabf47e234f4a1
Resolves: #55601
Reviewed-on: https://review.typo3.org/27330
Reviewed-by: Chris topher
Tested-by: Chris topher