[TASK] Update documentation of stdWrap of cObject TEXT
The property ".stdWrap" is only mentioned generally as part of the
"stdWrap functions" while it in contrast is explicitly mentioned for the
other cObjects. Document it explicitly here as well!
Update the note ("very rootlevel") accordingly; discourage usage of
those stdWrap properties, which are located directly on the rootlevel of
the cObject.

Remove one stdWrap property, which was listed although it does not exist
in that version of TYPO3.

Change-Id: I304b5e43caae29251a097e435ffabf47e234f4a1
Resolves: #55601
Reviewed-on: https://review.typo3.org/27331
Reviewed-by: Chris topher
Tested-by: Chris topher