[TASK] Add latexpdf settings to Settings.yml
With these settings a normal Sphinx-build for LaTeX-PDF is possible. Change into the ./Documentation/_make folder and issue the "make latexpdf" command. A system with Python, Sphinx, RestTools and Texlive is required. For setting up a Debian system in that way a Vagrantfile is under construction.

Exactly this build process should be possible this on the current docs.typo3.org server as well and produce ./Documentation/_make/build/latex/TyposcriptReference.pdf.

Q: Do these 'latex_documents' and 'latex_elements' settings interfere with the current build tools on docs.typo3.org?

They shouldn't. If they do we should garantee that on the server our build tools superseed these settings correctly.