[TASK] Switch default highlight lang to TS
Most code snippets in this manual are TypoScript. So it seems
a good idea to switch to 'typoscript' as default highlight
language. This should give most code snippets highlighting.

Some highlightings for PHP snippets may be gone. We need to
add a '.. highlight:: php' into the reST source in these
cases.