Enhance description how to find out menu item keys
The keys of the menu items in the info module may change over time,
so the docs will not be able to provide have current examples. Explain this
and explain a better way to find out which menu items use what kind of key.
Since menu items are registered, it is easier to use the configuration module
instead of inspecting html source code.