TASK: Improve information about addRootLineFields
Install Tool is no longer available with release of 9.x. Instead there
are Admin Tools and "Settings" module offering the "Global
Configuration".

Therefore we are more explicit and add the PHP Array syntax.

Also as it's now already mentioned, we do not need to repeat this
information within the example section anymore.