Add hint for 32bit operating systems
In `typo3/sysext/frontend/Classes/ContentObject/ContentObjectRenderer.php` `cObjGet()` filters and sorts the array keys with `filterAndSortByNumericKeys()` and the help of `MathUtility::canBeInterpretedAsInteger()`.

This is done by a type cast to `int` and a comparison with the original key. This type cast fails with numbers greater than `2147483647` on 32bit operation systems. Yes, there are still web hosting companies out there with 32bit operation systems!