Add example to getQueryParams
When using getQueryParam for query parameters generated by extensions
they will often be parameters with arrays. This may case error logs
in case the array (or the key) is not defined (e.g.
"Unable to get an item on a non-array".

To prevent this, the function traverse can be used.

This change adds another example using traverse() to getQueryParams()
for this common usecase.

Releases: master, 10.4, 9.5