Changing the order of the Content-supplying properties of stdWrap part (#793)
* Changing the order of the Content-supplying properties of stdWrap part

releases: main

Remark: In my opinion the "Example" and "Content-supplying properties" part, at the end of the site, should be moved back to the top, like in for example the 10.x version of TYPO3. Because if it is so far down, hardly anyone reads it, especially the part with the order is very important. If you see this differently or have suggestions for improvement through mistakes, then let me know.

https://docs.typo3.org/m/typo3/reference-typoscript/main/en-us/Functions/Stdwrap.html

* Changing the order of the Content-supplying properties of stdWrap part

releases: main

Remark: In my opinion the "Example" and "Content-supplying properties" part, at the end of the site, should be moved back to the top, like in for example the 10.x version of TYPO3. Because if it is so far down, hardly anyone reads it, especially the part with the order is very important. If you see this differently or have suggestions for improvement through mistakes, then let me know.

https://docs.typo3.org/m/typo3/reference-typoscript/main/en-us/Functions/Stdwrap.html