[Backport 11.5] [TASK] Split GIFBUILDER objects into separate pages (#823)
The GIFBUILDER objects are listed on one large page. This makes it
difficult to look out for a specific object. Therefore, this
"god" page is split up into separate pages - one for each object.
This way, the different objects are listed in the navigation and
are easily accessible by the reader.

Note: This change just splits up the page into multiple pages.
Further adjustments on the different objects may be done in
consecutive pull request.

Releases: main, 12.4, 11.5