[BUGFIX] Correct more references (#1107) (#1110)
Sphinx is here lax and looks, if there is a reference of that name in all of the
intersphinxes. But guides is more picky and throws a warning. Therefore, the
reference is now prefixed with the correct manual.

Additionally, the reference to "t3start:Templating" has been removed as this
reference is not available anymore and also not used in newer versions of this manual.

Releases: main, 12.4, 11.5