[TASK] Update property information (#1115)
* [TASK] Update property information

According to https://forge.typo3.org/issues/102315 it is no longer possible to skip the argument `extensionName` in the ViewHelpers f:translate and f:uri.resource with the property `extbase.controllerExtensionName` as of TYPO3 v12. For this reason, the use of this property should no longer be used as a tip in the documentation. Instead, a warning should be issued.

* Update Documentation/ContentObjects/Fluidtemplate/Index.rst

Co-authored-by: Garvin Hicking <blog@garv.in>

* Update Documentation/ContentObjects/Fluidtemplate/Index.rst

Co-authored-by: Garvin Hicking <blog@garv.in>

* [TASK] Update property information

---------

Co-authored-by: Chris Müller <2566282+brotkrueml@users.noreply.github.com>
Co-authored-by: Garvin Hicking <blog@garv.in>