[TASK] Split up page TSconfig's mod properties (#446)
The mod page is really long. By splitting up, the pages
are much smaller. Additionally, the properties are now visible in the main
navigation which eases the access.

In the new pages, only the header sections are adjusted, nothing else is done.
They can be streamlined in a follow-up.

Releases: main, 12.4, 11.5