[TASK] Move the TMENU as subpage of the HMENU
TMENUs are seldom used anymore since there data processors and
Templates are recommended. If they are used they are always used in the
context of a HMENU. So the current situation where there is a menu point
called "MENU" in the menu that doesnt tell you how to build a modern menu
is quite misleading.

No text changes made, so that Git can understand the files where moved.

Text changes will be applied in a follow-up to further distinguish the usage.

I will create redirects

Releases: main, 12.4, 11.5