[TASK] Move page.meta refresh example to dedicated section
The Open Graph meta tag examples are polluted by the meta refresh example.
This is now moved to the already available section.

The previous "[seconds]; [URL, leave blank for same page]" is removed
as this is not a manual about the "refresh" property but about setting
that property in TYPO3.

Releases: main, 12.4