[TASK] Move content of "Data Types"
There are no real data types in TypoScript, there are just certain
string formats expected in some properties.

Where the former "data type" is an expected input for exactly one function, move the expected input to that function.

Where it works like a function itself, move it to functions,

Move the rest of expected string formats to the syntax but make clear they are no real types.

Releases: main, 13.4