[BUGFIX] Display page and config on top level of Frontend TypoScript
The page is the starting point for whatever rendering is configured via TypoScript. It is hard to find within the "Top level object."

The TypoScript config also has a central and special handling.

Leave hints of how they can be found in the previous place in the Top Level objects

Releases: main, 13.4