[TASK] Display page and config on top level of Frontend TypoScript (#1525)
* [BUGFIX] Display page and config on top level of Frontend TypoScript

The page is the starting point for whatever rendering is configured via TypoScript. It is hard to find within the "Top level object."

The TypoScript config also has a central and special handling.

Leave hints of how they can be found in the previous place in the Top Level objects

Releases: main, 13.4

* Apply suggestions from code review

* Update Documentation/TopLevelObjects/PageAndConfig.rst

Co-authored-by: Stefan Frömken <froemken@gmail.com>

---------

Co-authored-by: Stefan Frömken <froemken@gmail.com>