[TASK] Unify extension name in examples (#1637)
* [TASK] Unify extension name in examples

Unify the extension name of examples to an identical in all places of documentation.

Follow-up of https://github.com/TYPO3-Documentation/TYPO3CMS-Reference-CoreApi/pull/6110

* Use backticks to mark this as code

Co-authored-by: Lina Wolf <48202465+linawolf@users.noreply.github.com>

* Re-add prefix to tablename

The prefix was previously removed due to an incorrect search & replace operation.

Co-authored-by: Lina Wolf <48202465+linawolf@users.noreply.github.com>

* Re-add exxtension prefix

The prefix was previously removed due to an incorrect search & replace operation.

Co-authored-by: Lina Wolf <48202465+linawolf@users.noreply.github.com>

* [BUGFIX] Revert incorrect replace action

In the previous commit, an incorrect search & replace action was performed.
This has now been reverted.

* [TASK] Unify extension name in examples

Unify the extension name of examples to an identical in all places of documentation.

Follow-up of https://github.com/TYPO3-Documentation/TYPO3CMS-Reference-CoreApi/pull/6110

---------

Co-authored-by: Lina Wolf <48202465+linawolf@users.noreply.github.com>